+++ sdk/keyvault/arm-keyvault/src/models/parameters.ts
@@ -120,7 +120,7 @@ export const keyName: OperationURLParameter = {
 export const apiVersion: OperationQueryParameter = {
   parameterPath: "apiVersion",
   mapper: {
-    defaultValue: "2024-11-01",
+    defaultValue: "2025-05-01",
     isConstant: true,
     serializedName: "api-version",
     type: {